refactor: update image with text section
